enable leak check only

+++ tests/unit_tests/args_with_s.sh
@@ -12,20 +12,25 @@ OPTION="-s"
 FAILED=0
 
 cd ${TEST_DATA_DIR}
-ls ${OPTION} testdir z textfile.txt . x symbolic_link_different_dir y &> ${EXPECTS_DIR}/${TEST_NAME}
-../../ft_ls ${OPTION} testdir z textfile.txt . x symbolic_link_different_dir y &> ${OUTPUTS_DIR}/${TEST_NAME}_output
-diff ${OUTPUTS_DIR}/${TEST_NAME}_output ${EXPECTS_DIR}/${TEST_NAME} > ${LOGS_DIR}/${TEST_NAME}.log
-if [ $? -eq 0 ]; then
-  rm ${LOGS_DIR}/${TEST_NAME}.log
-  echo -n "[PASS]"
-else
-  echo -n "-FAIL-"
-  FAILED=1
-fi
-echo " test: ${TEST_NAME}"
 
-if [ ${SKIP_LEAK_CHECK} ]; then
-  exit ${FAILED}
+if [ ${LEAK_CHECK_ONLY} ]; then
+  SKIP_LEAK_CHECK=
+else
+  ls ${OPTION} testdir z textfile.txt . x symbolic_link_different_dir y &> ${EXPECTS_DIR}/${TEST_NAME}
+  ../../ft_ls ${OPTION} testdir z textfile.txt . x symbolic_link_different_dir y &> ${OUTPUTS_DIR}/${TEST_NAME}_output
+  diff ${OUTPUTS_DIR}/${TEST_NAME}_output ${EXPECTS_DIR}/${TEST_NAME} > ${LOGS_DIR}/${TEST_NAME}.log
+  if [ $? -eq 0 ]; then
+    rm ${LOGS_DIR}/${TEST_NAME}.log
+    echo -n "[PASS]"
+  else
+    echo -n "-FAIL-"
+    FAILED=1
+  fi
+  echo " test: ${TEST_NAME}"
+
+  if [ ${SKIP_LEAK_CHECK} ]; then
+    exit ${FAILED}
+  fi
 fi
 
 valgrind --log-file="${LOGS_DIR}/${TEST_NAME}_leaks.log" --leak-check=full --error-exitcode=2 ../../ft_ls ${OPTION} testdir z textfile.txt . x symbolic_link_different_dir y &> /dev/null
